Fixing linter

+++ checkpoint/keeper.go
@@ -184,6 +184,7 @@ func (k *Keeper) GetLastCheckpoint(ctx sdk.Context) (hmTypes.Checkpoint, error) 
 			return _checkpoint, nil
 		}
 	}
+
 	return _checkpoint, cmn.ErrNoCheckpointFound(k.Codespace())
 }
 
@@ -238,8 +239,7 @@ func (k *Keeper) GetLastNoAck(ctx sdk.Context) uint64 {
 		// get current ACK count
 		result, err := strconv.ParseUint(string(store.Get(LastNoACKKey)), 10, 64)
 		if err == nil {
-
-			return uint64(result)
+			return result
 		}
 	}
 	return 0
@@ -262,6 +262,7 @@ func (k *Keeper) GetCheckpoints(ctx sdk.Context) []hmTypes.Checkpoint {
 			headers = append(headers, checkpoint)
 		}
 	}
+
 	return headers
 }
